+++ CHANGELOG.md
@@ -11,3 +11,9 @@
 ## 0.10.1-hotfix.1
 - fix async camera moving issue on Android
 - change to kotlin code style (Android Native)
+
+## 0.10.2
+- fix moveCamera issue (using CameraUpdate.scrollTo method)
+- delete moveCamera enableAnimation option
+- add moveCamera animationDuration option
+- change to kotlin code style (Android Native)